requirements.txt
asgiref==3.8.1
diff-match-patch==20230430
Django==5.1.1
django-import-export==4.1.1
djangorestframework==3.15.2
et-xmlfile==1.1.0
openpyxl==3.1.5
psycopg==3.2.1
sqlparse==0.5.1
tablib==3.5.0
typing_extensions==4.12.2
xlrd==2.0.1
xlwt==1.3.0
python-dotenv==1.0.1